separated split quote in lexer functions
